// frontend/types/generated/room.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { RoomId } from "./branded_types";
import type { UserId } from "./branded_types";

export type Room = { id: RoomId, slug: string, name: string, owner_id: UserId, created_at: string, updated_at: string, };